Moved files from /reformulation to /ingredients/constraint_relaxation_strategies, and from /solvers to /ingredients/subproblem_solvers
