+++ uno/ingredients/hessian_models/HessianModelFactory.cpp
@@ -6,7 +6,7 @@
 #include "ConvexifiedHessian.hpp"
 #include "ExactHessian.hpp"
 #include "ZeroHessian.hpp"
-#include "solvers/DirectSymmetricIndefiniteLinearSolver.hpp"
+#include "ingredients/subproblem_solvers/DirectSymmetricIndefiniteLinearSolver.hpp"
 
 namespace uno {
    std::unique_ptr<HessianModel> HessianModelFactory::create(const std::string& hessian_model, size_t dimension, size_t maximum_number_nonzeros,
