+++ uno/ingredients/inequality_handling_methods/inequality_constrained_methods/InequalityConstrainedMethod.cpp
@@ -5,7 +5,7 @@
 #include "optimization/Direction.hpp"
 #include "optimization/Iterate.hpp"
 #include "linear_algebra/Vector.hpp"
-#include "reformulation/l1RelaxedProblem.hpp"
+#include "ingredients/constraint_relaxation_strategies/l1RelaxedProblem.hpp"
 #include "options/Options.hpp"
 #include "symbolic/VectorView.hpp"
