+++ uno/ingredients/inequality_handling_methods/inequality_constrained_methods/LPSubproblem.cpp
@@ -4,9 +4,9 @@
 #include "LPSubproblem.hpp"
 #include "optimization/Direction.hpp"
 #include "optimization/WarmstartInformation.hpp"
-#include "reformulation/OptimizationProblem.hpp"
-#include "solvers/LPSolver.hpp"
-#include "solvers/LPSolverFactory.hpp"
+#include "ingredients/constraint_relaxation_strategies/OptimizationProblem.hpp"
+#include "ingredients/subproblem_solvers/LPSolver.hpp"
+#include "ingredients/subproblem_solvers/LPSolverFactory.hpp"
 #include "options/Options.hpp"
 
 namespace uno {
