+++ uno/ingredients/inequality_handling_methods/inequality_constrained_methods/QPSubproblem.cpp
@@ -3,15 +3,13 @@
 
 #include "QPSubproblem.hpp"
 #include "optimization/Direction.hpp"
-#include "ingredients/hessian_models/HessianModelFactory.hpp"
-#include "linear_algebra/SymmetricMatrix.hpp"
 #include "optimization/Iterate.hpp"
 #include "optimization/WarmstartInformation.hpp"
 #include "preprocessing/Preprocessing.hpp"
-#include "reformulation/OptimizationProblem.hpp"
-#include "solvers/DirectSymmetricIndefiniteLinearSolver.hpp"
-#include "solvers/QPSolver.hpp"
-#include "solvers/QPSolverFactory.hpp"
+#include "ingredients/constraint_relaxation_strategies/OptimizationProblem.hpp"
+#include "ingredients/subproblem_solvers/DirectSymmetricIndefiniteLinearSolver.hpp"
+#include "ingredients/subproblem_solvers/QPSolver.hpp"
+#include "ingredients/subproblem_solvers/QPSolverFactory.hpp"
 #include "options/Options.hpp"
 #include "tools/Statistics.hpp"
