+++ uno/ingredients/subproblem_solvers/MA27/MA27Solver.hpp
@@ -6,7 +6,7 @@
 
 #include <array>
 #include <vector>
-#include "solvers/DirectSymmetricIndefiniteLinearSolver.hpp"
+#include "../DirectSymmetricIndefiniteLinearSolver.hpp"
 
 namespace uno {
    // forward declaration
