+++ uno/linear_algebra/SymmetricIndefiniteLinearSystem.hpp
@@ -9,11 +9,10 @@
 #include "SparseStorageFactory.hpp"
 #include "RectangularMatrix.hpp"
 #include "ingredients/hessian_models/UnstableRegularization.hpp"
+#include "ingredients/subproblem_solvers/DirectSymmetricIndefiniteLinearSolver.hpp"
 #include "model/Model.hpp"
 #include "optimization/WarmstartInformation.hpp"
-#include "solvers/DirectSymmetricIndefiniteLinearSolver.hpp"
 #include "options/Options.hpp"
-#include "solvers/DirectSymmetricIndefiniteLinearSolver.hpp"
 #include "tools/Statistics.hpp"
 
 namespace uno {
