+++ unotest/functional_tests/HiGHSSolverTests.cpp
@@ -2,12 +2,12 @@
 // Licensed under the MIT license. See LICENSE file in the project directory for details.
 
 #include <gtest/gtest.h>
-#include "optimization/Direction.hpp"
+#include "ingredients/subproblem_solvers/HiGHS/HiGHSSolver.hpp"
 #include "linear_algebra/RectangularMatrix.hpp"
 #include "linear_algebra/SparseVector.hpp"
+#include "optimization/Direction.hpp"
 #include "optimization/WarmstartInformation.hpp"
 #include "options/Options.hpp"
-#include "solvers/HiGHS/HiGHSSolver.hpp"
 #include "tools/Infinity.hpp"
 
 using namespace uno;
