+++ unotest/functional_tests/MA57SolverTests.cpp
@@ -2,8 +2,8 @@
 // Licensed under the MIT license. See LICENSE file in the project directory for details.
 
 #include <gtest/gtest.h>
+#include "ingredients/subproblem_solvers/MA57/MA57Solver.hpp"
 #include "linear_algebra/SymmetricMatrix.hpp"
-#include "solvers/MA57/MA57Solver.hpp"
 
 using namespace uno;
