+++ unotest/functional_tests/MUMPSSolverTests.cpp
@@ -2,8 +2,8 @@
 // Licensed under the MIT license. See LICENSE file in the project directory for details.
 
 #include <gtest/gtest.h>
+#include "ingredients/subproblem_solvers/MUMPS/MUMPSSolver.hpp"
 #include "linear_algebra/SymmetricMatrix.hpp"
-#include "solvers/MUMPS/MUMPSSolver.hpp"
 
 using namespace uno;
